move gardener client and hyperscaler package code to common directory
