+++ components/kyma-environment-broker/Dockerfile.sac
@@ -4,6 +4,7 @@ FROM golang:1.14.4-alpine3.12 AS build
 WORKDIR /go/src/github.com/kyma-project/control-plane/components/kyma-environment-broker
 
 COPY cmd cmd
+COPY common common
 COPY internal internal
 COPY vendor vendor
