+++ components/kyma-environment-broker/common/hyperscaler/azure/config.go
