+++ components/kyma-environment-broker/common/hyperscaler/azure/env.go
@@ -5,9 +5,9 @@ import (
 
 	"github.com/Azure/go-autorest/autorest/azure"
 
+	"github.com/kyma-project/control-plane/components/kyma-environment-broker/common/hyperscaler"
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal"
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/broker"
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/hyperscaler"
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/provider"
 )
