+++ components/kyma-environment-broker/common/hyperscaler/azure/provider.go
