+++ components/kyma-environment-broker/common/hyperscaler/shared_pool.go
