+++ components/kyma-environment-broker/internal/process/deprovisioning/event_hub_step.go
@@ -7,9 +7,9 @@ import (
 
 	"github.com/sirupsen/logrus"
 
+	"github.com/kyma-project/control-plane/components/kyma-environment-broker/common/hyperscaler"
+	"github.com/kyma-project/control-plane/components/kyma-environment-broker/common/hyperscaler/azure"
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal"
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/hyperscaler"
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/hyperscaler/azure"
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/process"
 	processazure "github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/process/azure"
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/storage"
