+++ components/kyma-environment-broker/internal/process/provisioning/event_hub_step.go
@@ -11,9 +11,9 @@ import (
 	"github.com/Azure/azure-sdk-for-go/services/eventhub/mgmt/2017-04-01/eventhub"
 	"github.com/sirupsen/logrus"
 
+	"github.com/kyma-project/control-plane/components/kyma-environment-broker/common/hyperscaler"
+	"github.com/kyma-project/control-plane/components/kyma-environment-broker/common/hyperscaler/azure"
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal"
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/hyperscaler"
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/hyperscaler/azure"
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/process"
 	processazure "github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/process/azure"
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/ptr"
